+++ test/resources/language/c_c++/header.h
@@ -1,0 +1,5 @@
+#pragma once
+
+static inline int flycheck_cmp(int a, unsigned b) {
+  return a < b;
+}